+++ test/aoj/GRL_1_C.test.cpp
@@ -5,19 +5,21 @@
 int main() {
     size_t v, e;
     cin >> v >> e;
-    using P = pair<int, i64>;
+    using P = pair<size_t, long long>;
     vector<vector<P>> graph(v, vector<P>{});
     while (e--) {
-        int s, t;
-        i64 d;
+        size_t s, t;
+        long long d;
         cin >> s >> t >> d;
-        graph[s].emplace_back(P{t, d});
+        graph[s].emplace_back(P{ t, d });
     }
     const auto result = floyd_warshall(graph);
-    if (empty(result)) cout << "NEGATIVE CYCLE\n";
+    if (empty(result))
+        cout << "NEGATIVE CYCLE\n";
     else
         rep(i, v) rep(j, v) {
-            if (result[i][j] >= numeric_limits<i64>::max()) cout << "INF";
+            if (result[i][j] >= numeric_limits<long long>::max())
+                cout << "INF";
             else
                 cout << result[i][j];
             cout << " \n"[j + 1 == v];
