+++ test/library-checker/lca.test.cpp
@@ -3,17 +3,17 @@
 #include "lib/graph/lowest_common_ancestor.hpp"
 
 int main() {
-    int n, q;
+    size_t n, q;
     cin >> n >> q;
-    vector graph(n, vector<int>{});
+    vector graph(n, vector<size_t>{});
     rep(i, 1, n) {
-        int p;
+        size_t p;
         cin >> p;
         graph[p].push_back(i);
     }
     LCA lca(graph, 0);
     while (q--) {
-        int u, v;
+        size_t u, v;
         cin >> u >> v;
         cout << lca(u, v) << "\n";
     }
